# openapi.yaml
openapi: 3.0.3
info:
  title: Blockvisor API
  description: |
    Blockvisor API is a distributed infrastructure management system that orchestrates software deployment and monitoring across physical and virtual machines. Originally designed for blockchain node management, it has evolved into a general-purpose platform for running and managing any software.
    
    ## Core Concepts
    - **Hosts**: Physical or virtual machines registered with the system via the `bvup` tool
    - **Nodes**: Running software instances with real-time state and metrics monitoring
    - **Organizations**: Groups of users sharing resources and permissions
    - **Users**: Individual accounts with role-based access control (RBAC)
    - **Commands**: Instructions sent to hosts for software lifecycle management
    
    ## Authentication
    The API uses JWT tokens for authentication. Include the token in the `Authorization` header as `Bearer <token>`.
    
    ## Error Handling
    The API returns structured error responses with detailed parameter validation information:
    ```json
    {
      "error": "Invalid query parameters",
      "details": [
        {
          "parameter": "org_id",
          "error": "Invalid UUID format: 'not-a-uuid'",
          "expected": "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"
        }
      ]
    }
    ```
  version: 0.4.0
  contact:
    name: Blockvisor API Support
  license:
    name: MIT
servers:
  - url: https://api.blockvisor.com
...
security:
  - BearerAuth: []
paths:
...
  /v1/node:
    get:
      tags:
        - Nodes
      summary: List nodes
      description: Get a list of nodes with optional filtering
      parameters:
        - name: org_id
          in: query
          description: Organization ID to filter by (singular form)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: org_ids
          in: query
          description: Organization IDs to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000,6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        - name: host_id
          in: query
          description: Host ID to filter by (singular form)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: host_ids
          in: query
          description: Host IDs to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000,6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        - name: user_id
          in: query
          description: User ID to filter by (singular form)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: user_ids
          in: query
          description: User IDs to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000,6ba7b810-9dad-11d1-80b4-00c04fd430c8"
        - name: protocol_ids
          in: query
          description: Protocol IDs to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "protocol1,protocol2"
        - name: semantic_versions
          in: query
          description: Semantic versions to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "1.0.0,1.1.0"
        - name: ip_addresses
          in: query
          description: IP addresses to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "192.168.1.1,10.0.0.1"
        - name: node_states
          in: query
          description: Node states to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "running,stopped"
            enum: [starting, running, stopped, failed, upgrading, deleting, deleted]
        - name: next_states
          in: query
          description: Next states to filter by (comma-separated or repeated)
          schema:
            type: string
            example: "stopping,upgrading"
            enum: [stopping, deleting, upgrading]
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
            example: 50
      responses:
        '200':
          description: Nodes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'
                  total:
                    type: integer
                    description: Total number of nodes
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    post:
      tags:
        - Nodes
      summary: Create node
      description: Create a new node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeRequest'
      responses:
        '201':
          description: Node created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  node:
                    $ref: '#/components/schemas/Node'
  /v1/node/{id}:
    get:
      tags:
        - Nodes
      summary: Get node
      description: Get details of a specific node
      parameters:
        - name: id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Node retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  node:
                    $ref: '#/components/schemas/Node'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Nodes
      summary: Delete node
      description: Delete a node
      parameters:
        - name: id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Node deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Node deleted successfully"
  /v1/node/{id}/start:
    put:
      tags:
        - Nodes
      summary: Start node
      description: Start a stopped node
      parameters:
        - name: id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Node start command sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Node start command sent"
  /v1/node/{id}/stop:
    put:
      tags:
        - Nodes
      summary: Stop node
      description: Stop a running node
      parameters:
        - name: id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Node stop command sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Node stop command sent"
  /v1/node/{id}/restart:
    put:
      tags:
        - Nodes
      summary: Restart node
      description: Restart a node
      parameters:
        - name: id
          in: path
          required: true
          description: Node ID
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Node restart command sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Node restart command sent"

  # Host endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred"
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid query parameters"
        details:
          type: array
          items:
            type: object
            properties:
              parameter:
                type: string
                example: "org_id"
              error:
                type: string
                example: "Invalid UUID format: 'not-a-uuid'"
              expected:
                type: string
                example: "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"

    # Core entity schemas
...
    Node:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        org_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        host_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        protocol_id:
          type: string
          example: "bitcoin"
        state:
          type: string
          enum: [starting, running, stopped, failed, upgrading, deleting, deleted]
          example: "running"
        ip_address:
          type: string
          format: ipv4
          example: "192.168.1.100"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
...
    CreateNodeRequest:
      type: object
      required:
        - org_id
        - host_id
        - protocol_id
      properties:
        org_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        host_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        protocol_id:
          type: string
          example: "bitcoin"
        version:
          type: string
          example: "1.0.0"
        config:
          type: object
          description: Protocol-specific configuration
          additionalProperties: true
...